Split Cholesky off into module
