+++ src/matrix/decomposition/mod.rs
@@ -11,7 +11,7 @@
 //! (http://www.cs.utexas.edu/users/inderjit/public_papers/HLA_SVD.pdf)
 
 mod qr;
-
+mod cholesky;
 
 use std::any::Any;
 use std::cmp;
@@ -30,70 +30,6 @@ use libnum::{cast, abs};
 impl<T> Matrix<T>
     where T: Any + Float
 {
-    /// Cholesky decomposition
-    ///
-    /// Returns the cholesky decomposition of a positive definite matrix.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use rulinalg::matrix::Matrix;
-    ///
-    /// let m = Matrix::new(3,3, vec![1.0,0.5,0.5,0.5,1.0,0.5,0.5,0.5,1.0]);
-    ///
-    /// let l = m.cholesky();
-    /// ```
-    ///
-    /// # Panics
-    ///
-    /// - The matrix is not square.
-    ///
-    /// # Failures
-    ///
-    /// - Matrix is not positive definite.
-    pub fn cholesky(&self) -> Result<Matrix<T>, Error> {
-        assert!(self.rows == self.cols,
-                "Matrix must be square for Cholesky decomposition.");
-
-        let mut new_data = Vec::<T>::with_capacity(self.rows() * self.cols());
-
-        for i in 0..self.rows() {
-
-            for j in 0..self.cols() {
-
-                if j > i {
-                    new_data.push(T::zero());
-                    continue;
-                }
-
-                let mut sum = T::zero();
-                for k in 0..j {
-                    sum = sum + (new_data[i * self.cols() + k] * new_data[j * self.cols() + k]);
-                }
-
-                if j == i {
-                    new_data.push((self[[i, i]] - sum).sqrt());
-                } else {
-                    let p = (self[[i, j]] - sum) / new_data[j * self.cols + j];
-
-                    if !p.is_finite() {
-                        return Err(Error::new(ErrorKind::DecompFailure,
-                                              "Matrix is not positive definite."));
-                    } else {
-
-                    }
-                    new_data.push(p);
-                }
-            }
-        }
-
-        Ok(Matrix {
-            rows: self.rows(),
-            cols: self.cols(),
-            data: new_data,
-        })
-    }
-
     /// Compute the cos and sin values for the givens rotation.
     ///
     /// Returns a tuple (c, s).
@@ -1394,14 +1330,6 @@ mod tests {
         assert!(eigs.iter().any(|x| (x - eig_3).abs() < 1e-4));
         assert!(eigs.iter().any(|x| (x - eig_4).abs() < 1e-4));
         assert!(eigs.iter().any(|x| (x - eig_5).abs() < 1e-4));
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_non_square_cholesky() {
-        let a = Matrix::new(2, 3, vec![1.0; 6]);
-
-        let _ = a.cholesky();
     }
 
     #[test]
